ARM: zynq: HACK: Define environment to be on SD card
This change hard codes the environment to be on the SD card just behind
the MBR.  In contrast to the standard setting of the environment to be
in QSPI, this allows an SD card to be completely self containing.
Moving the SD card between systems will correctly carry the environment
to the target system.

Signed-off-by: Detlev Zundel <detlev.zundel@ebv.com>